GenAI_SourceCode/16saporchastrate/requirements.txt
langchain
langchain-community
langchain-openai
langchain-core
sap-ai-sdk-core
sap-ai-sdk-gen
cfenv
langserve
python-dotenv
sse_starlette
streamlit
uvicorn